backend/requirements.txt
# BookFlow Backend - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & Storage
supabase==2.3.4

# PDF Processing
PyMuPDF==1.23.8
weasyprint==60.2

# AI
anthropic==0.18.1

# HTML Processing
beautifulsoup4==4.12.3
Jinja2==3.1.3

# Auth
PyJWT==2.8.0
cryptography==42.0.2

# Utils
python-dotenv==1.0.0
httpx>=0.24.0,<0.26.0
